pyproject.toml
[project]
name = "excelclaw-whatsapp"
version = "0.1.0"
description = "ExcelClaw WhatsApp Bot — Turn your Excel into an AI agent via WhatsApp"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --port 8000"
